refactor(query): extract QueryMixin composed into BaseGnuCashBook
Move _query_filtered_splits from ReportingMixin into a new
book/_query.py / QueryMixin, inherited by BaseGnuCashBook
unconditionally — same pattern as CurrencyMixin (Stage 1). Pure
motion; tests at 1,133 passing.

The helper is a query primitive that multiple modules need
(reporting today, budgets next). Living on ReportingMixin meant
budgets couldn't use it without reporting also loaded. Promoting
to BaseGnuCashBook removes the implicit reporting dependency.

Unblocks the next commit: pushing get_budget_report filtering
to SQL.
